move Project module out of peer
